+++ libp2p/src/lib.rs
@@ -133,15 +133,15 @@ pub use libp2p_websocket as websocket;
 #[cfg(all(feature = "websocket-websys", target_arch = "wasm32"))]
 #[doc(inline)]
 pub use libp2p_websocket_websys as websocket_websys;
-#[cfg(all(feature = "webtransport-websys", target_arch = "wasm32"))]
-#[cfg_attr(docsrs, doc(cfg(feature = "webtransport-websys")))]
-#[doc(inline)]
-pub use libp2p_webtransport_websys as webtransport_websys;
 #[cfg(feature = "webtransport")]
 #[cfg(not(target_arch = "wasm32"))]
 #[cfg_attr(docsrs, doc(cfg(feature = "webtransport")))]
 #[doc(inline)]
 pub use libp2p_webtransport as webtransport;
+#[cfg(all(feature = "webtransport-websys", target_arch = "wasm32"))]
+#[cfg_attr(docsrs, doc(cfg(feature = "webtransport-websys")))]
+#[doc(inline)]
+pub use libp2p_webtransport_websys as webtransport_websys;
 #[cfg(feature = "yamux")]
 #[doc(inline)]
 pub use libp2p_yamux as yamux;